+++ source/solvers/tracer_scratch_data.cc
@@ -28,11 +28,13 @@ TracerScratchData<dim>::allocate()
   this->tracer_gradients = std::vector<Tensor<1, dim>>(n_q_points);
   this->previous_tracer_gradients =
     std::vector<Tensor<1, dim>>(this->n_q_points);
-  this->tracer_laplacians       = std::vector<double>(n_q_points);
-  this->tracer_diffusivity      = std::vector<double>(n_q_points);
-  this->tracer_diffusivity_face = std::vector<double>(n_q_points);
-  this->tracer_diffusivity_0    = std::vector<double>(n_q_points);
-  this->tracer_diffusivity_1    = std::vector<double>(n_q_points);
+  this->tracer_laplacians              = std::vector<double>(n_q_points);
+  this->tracer_diffusivity             = std::vector<double>(n_q_points);
+  this->tracer_diffusivity_face        = std::vector<double>(n_q_points);
+  this->tracer_diffusivity_0           = std::vector<double>(n_q_points);
+  this->tracer_diffusivity_1           = std::vector<double>(n_q_points);
+  this->tracer_reaction_prefactor      = std::vector<double>(n_q_points);
+  this->grad_tracer_reaction_prefactor = std::vector<double>(n_q_points);
 
   // Solid signed distance function
   if (properties_manager.field_is_required(field::levelset))
@@ -57,6 +59,10 @@ TracerScratchData<dim>::allocate()
   if (properties_manager.field_is_required(field::levelset))
     fields.insert(
       std::pair<field, std::vector<double>>(field::levelset, n_q_points));
+  if (properties_manager.field_is_required(field::tracer_concentration))
+    fields.insert(
+      std::pair<field, std::vector<double>>(field::tracer_concentration,
+                                            n_q_points));
 }
 
 
@@ -67,14 +73,27 @@ TracerScratchData<dim>::calculate_physical_properties()
   if (properties_manager.field_is_required(field::levelset))
     set_field_vector(field::levelset, this->sdf_values, this->fields);
 
+  if (properties_manager.field_is_required(field::tracer_concentration))
+    set_field_vector(field::tracer_concentration,
+                     this->tracer_values,
+                     this->fields);
+
   switch (properties_manager.get_number_of_fluids())
     {
       case 1:
         {
-          // In this case, only viscosity is the required property
           const auto diffusivity_model =
             properties_manager.get_tracer_diffusivity();
           diffusivity_model->vector_value(fields, tracer_diffusivity);
+          const auto reaction_prefactor_model =
+            properties_manager.get_tracer_reaction_prefactor();
+          reaction_prefactor_model->vector_value(fields,
+                                                 tracer_reaction_prefactor);
+          reaction_prefactor_model->vector_jacobian(
+            fields,
+            field::tracer_concentration,
+            grad_tracer_reaction_prefactor);
+
           break;
         }
       case 2: